% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results.R
\name{collectExperimentResults}
\alias{collectExperimentResults}
\title{Collect completed experiment results}
\usage{
collectExperimentResults(outputFolder)
}
\arguments{
\item{outputFolder}{Experiment output folder.}
}
\value{
List containing job status, metrics, and paired confidence intervals.
}
\description{
Collect completed experiment results
}
