% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results.R
\name{plotLearningCurves}
\alias{plotLearningCurves}
\title{Plot paired learning curves}
\usage{
plotLearningCurves(
  results,
  metric = "auroc",
  abscissa = c("events", "patients")
)
}
\arguments{
\item{results}{Output of collectExperimentResults.}

\item{metric}{One of auroc, auprc, logLoss, brier, calibrationIntercept, calibrationSlope.}

\item{abscissa}{Training events (default) or patients.}
}
\value{
A ggplot showing individual repetitions and their mean.
}
\description{
Plot paired learning curves
}
